amdvlk: Attempt to redefine function
root cause: Per SPIRV spec, Depth is not a property of type, we should not include it in type mangle name.

Following is from SPIRV spec OpTypeImage
Depth is whether or not this image is a depth image. (Note that whether or not depth comparisons are actually done is a property of the sampling opcode, not of this type declaration.)

ReviewBoardURL = http://dxreview2.amd.com/r/6271/diff/