issue#2: Mismatch of OpName on entry point and OpEntryPoint name causes crash
Port internal fix to dev branch

Root cause: The name of entry point may be different with the function name. and we should not use the function name in translator, it is debug info.

Solution: per SPIR-V spec, entry point name + exec model are unique in shader module, and translate need query entry point via function Id frequently, it is not easy support fast search with a set/map.
consider there are not a lot of entry point in a shader module, a simple vector is used to simplify the implementation.

Change:
1. Replace EntryPointSet and EntryPointVec with EntryPointVec in SPIRVModuleImpl
2. Add getExecModel and getName in SPIRVEntryPoint
3. Replace isEntryPoint with getEntryPoint, and support query entry point via function Id or entry point name + exec model.
4. Modify related functions to match the change in interface and class member.