Fix 58 failures of vulkan-cts-1.1.0.3
We have a shader with an OpAccessChain into a vector of
bool. SPIRVReader was converting that to a GEP into a vector of i1,
which is not supported by LLVM; see https://bugs.llvm.org/show_bug.cgi?id=1784

This commit changes SPIRVReader to use i8 instead of i1 for bool or
vector of bool when it appears in memory, or in a struct or array
(incuding matrix). LLVM optimization passes remove the i1<->i8
conversions in all reasonable cases.

Affects: dEQP-VK.glsl.arrays.*