Change to unrolling options
Modified the default unrolling options that are passed to LLVM. This
change mitigates the effects of disabling the loop unrolling guidance
pass in LLPC.

Between the two changes, the overall impact is expected to be fairly
neutral, with performance improving in some cases and degrading in
others.

This will form a better baseline for further unrolling changes.