Tuning option for loop unroll threshold
Add a shader tuning option "unrollThreshold" to allow the default loop unroll
threshold used by LLVM to be initialised to the specified value.
If a non-zero unrollThreshold is specified then it is passed to LLVM by use
of an attribute "amdgpu-unroll-threshold" that is attached to each function.
The unroll threshold used for each loop is then initialised to the specified
value before the target specific heuristics are applied, which may increase it.