Allow for llpc directory move: attempt 2
We are about to move most of the contents of the root of the llpc
repository into an llpc subdirectory.

This commit prepares the xgl CMakeLists.txt file for that move. This
time I added FORCE so it does not remember the old setting of
XGL_LLPC_PATH, and I changed it to detect the presence of the llpc
CMakeLists.txt, in case an empty llpc subdirectory is left over from the
last build.

Change-Id: I52c6c3de397ea756258912e760f533c1abfbe846