Capture SQTT for requested number of frames
The driver would capture the number of frames set up through
amdPalSettings.cfg` even when a specific number of frames is requested
through gpuopen's RGPProtocol via index mode. This fixes that issue and
captures the correct number of frames.

The .rgp file also contains incorrect information about the range
of frames captured via index mode in the API_INFO chunk. This is also
fixed so that information is reported correctly.