Allow to disable internal on-disk pipeline caching
Add a new panel setting, `EnableInternalPipelineCachingToDisk` that
controlls whether the ICD will write out its internal pipeline caches to
disk.

Note that this sounds similar to the existing setting
`AllowVkPipelineCachingToDisk`, but the key difference is that this
existing one controlls whether an internal cache based on PBC will be
created at all, while the new setting from this PR allows for internal
PBC caches that are not backed by archive files.