In PipelineCache, return nullptr cache adapter if PBC is nullptr
In the PipelineCache class, the member `m_pBinaryCache` will be null if
the `allowExternalPipelineCacheObject` setting is false. See
PipelineCache::Create.  However, the function `PipelineCache::GetCacheAdapter`
assumes that `m_pBinaryCache` is not nullptr.  This can cause a
segmentation fault.

This is fixed by returning `nullptr` in `PipelineCache::GetCacheAdapter`
if `m_pBinaryCache` is nullptr.