Remove "-add-hash-to-elf" usage from tests
This option is no longer needed by the test (since the hash is fetched
from PAL metadata), and will be removed from LLPC in a follow-up.

The small SPIR-V change is to prevent the ELF from exactly matching the
ELF of RunLitCommands.spvasm. If they are the same, then they hash the
the same MD5 sum and the correct "matched source file" cannot be
determined.