Write immutable sampler descriptors to descriptor sets
When compiling a relcoatable shader, it will not have the pipeline
layout, so it will not be able to embed the value for the immutable
sampler into the shader.  The code will try to load the sampler from the
descriptor set.

This change will have XGL write the immutable samplers when relocatable
shader are used.

This implements what was suggested in #129.