Fix writing immutable combined sampler
When writing the immutable samples so they can be read as a normal
sampler, it is written to the start of the descriptor.  However, if the
sampler is part of a combined image&sampler, it needs to be written
after the image descriptor.  This change will add that offset when
necessary.