Avoid unaligned memory accesses in PBC deserialization
The pipeline binary cache (PBC) data format is as follows:
```
| Public Header | Private Header (20B) | Entry (24B) | Blob (n) | Entry (24B) | ...
```

Binary cache entry headers (`Entry` above) require the alignment of 8 bytes,
which is not guaranteed with this data format. Therefore we cannot use
`reinterpret_cast` to read cache entry headers when reading PBC blobs.

This bug was found with the undefined behavior sanitizer. Although we have not
seen this being a problem in real-world applications, we would like to
get this fixed so that we run the code under UBSan and find other issues
in the future.

Fixes: https://github.com/GPUOpen-Drivers/xgl/issues/143.