Handling writing combined image/static sampler arrays
For relocatable shader compilation, we want to be able to write static
samplers to memory so that the compiler can treat the static samplers
the same as regular samplers.

The commit fixes the cases where the static sampler is part of an array
of combined image/sampler.  In this case, the samplers are not
contiguous in memory, as the current code assumes.  We need to identify
this case, and write the samplers in the correct locations.